backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.2",
    "langfuse>=3.5.0",
    "llama-index>=0.14.2",
    "openinference-instrumentation-llama-index>=4.3.5",
    "resend>=2.14.0",
    "supabase>=2.19.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
